+++ apollo-router-core/src/schema.rs
@@ -7,10 +7,10 @@ pub struct Schema {
     string: String,
     subtype_map: HashMap<String, HashSet<String>>,
     subgraphs: HashMap<String, String>,
-    object_types: HashMap<String, ObjectType>,
+    pub(crate) object_types: HashMap<String, ObjectType>,
     interfaces: HashMap<String, Interface>,
-    custom_scalars: HashSet<String>,
-    fragments: Fragments,
+    pub(crate) custom_scalars: HashSet<String>,
+    pub(crate) fragments: Fragments,
 }
 
 impl std::str::FromStr for Schema {
@@ -246,27 +246,26 @@ impl std::str::FromStr for Schema {
 }
 
 impl Schema {
+    /// TODO
     pub fn read(path: impl AsRef<std::path::Path>) -> Result<Self, SchemaError> {
         std::fs::read_to_string(path)?.parse()
     }
 
+    /// TODO
     pub fn as_str(&self) -> &str {
         &self.string
     }
 
-    pub fn is_subtype(&self, abstract_type: &str, maybe_subtype: &str) -> bool {
+    pub(crate) fn is_subtype(&self, abstract_type: &str, maybe_subtype: &str) -> bool {
         self.subtype_map
             .get(abstract_type)
             .map(|x| x.contains(maybe_subtype))
             .unwrap_or(false)
     }
 
+    /// TODO
     pub fn subgraphs(&self) -> impl Iterator<Item = (&String, &String)> {
         self.subgraphs.iter()
-    }
-
-    pub(crate) fn fragments(&self) -> &Fragments {
-        &self.fragments
     }
 }
 
@@ -280,7 +279,7 @@ macro_rules! implement_object_type_or_interface {
         }
 
         impl $name {
-            fn validate_object(&self, object: &Object, schema: &Schema) -> bool {
+            pub(crate) fn validate_object(&self, object: &Object, schema: &Schema) -> bool {
                 self
                     .fields
                     .iter()
@@ -361,149 +360,6 @@ implement_object_type_or_interface!(
     ast::InterfaceTypeDefinition,
     ast::InterfaceTypeExtension,
 );
-
-// Primitives are taken from scalars: https://spec.graphql.org/draft/#sec-Scalars
-#[derive(Debug)]
-pub(crate) enum FieldType {
-    Named(String),
-    List(Box<FieldType>),
-    NonNull(Box<FieldType>),
-    String,
-    Int,
-    Float,
-    Id,
-    Boolean,
-}
-
-impl FieldType {
-    pub(crate) fn validate_value(&self, value: &Value, schema: &Schema) -> bool {
-        match (self, value) {
-            (FieldType::String, Value::String(_)) => true,
-            (FieldType::Int, Value::Number(number)) if !number.is_f64() => true,
-            (FieldType::Float, Value::Number(number)) if number.is_f64() => true,
-            // "The ID scalar type represents a unique identifier, often used to refetch an object
-            // or as the key for a cache. The ID type is serialized in the same way as a String;
-            // however, it is not intended to be human-readable. While it is often numeric, it
-            // should always serialize as a String."
-            //
-            // In practice it seems Int works too
-            (FieldType::Id, Value::String(_) | Value::Number(_)) => true,
-            (FieldType::Boolean, Value::Bool(_)) => true,
-            (FieldType::List(inner_ty), Value::Array(vec)) => {
-                vec.iter().all(|x| inner_ty.validate_value(x, schema))
-            }
-            (FieldType::NonNull(inner_ty), value) => {
-                !value.is_null() && inner_ty.validate_value(value, schema)
-            }
-            (FieldType::Named(name), _) if schema.custom_scalars.contains(name) => true,
-            (FieldType::Named(name), value) if value.is_object() => {
-                if let Some(object_ty) = schema.object_types.get(name) {
-                    object_ty.validate_object(value.as_object().unwrap(), schema)
-                } else {
-                    false
-                }
-            }
-            // NOTE: graphql's types are all optional by default
-            // TODO add integration test for this
-            (_, Value::Null) => true,
-            _ => false,
-        }
-    }
-}
-
-impl From<ast::Type> for FieldType {
-    // Spec: https://spec.graphql.org/draft/#sec-Type-References
-    fn from(ty: ast::Type) -> Self {
-        match ty {
-            ast::Type::NamedType(named) => named.into(),
-            ast::Type::ListType(list) => list.into(),
-            ast::Type::NonNullType(non_null) => non_null.into(),
-        }
-    }
-}
-
-impl From<ast::NamedType> for FieldType {
-    // Spec: https://spec.graphql.org/draft/#NamedType
-    fn from(named: ast::NamedType) -> Self {
-        let name = named
-            .name()
-            .expect("the node Name is not optional in the spec; qed")
-            .text()
-            .to_string();
-        match name.as_str() {
-            "String" => Self::String,
-            "Int" => Self::Int,
-            "Float" => Self::Float,
-            "ID" => Self::Id,
-            "Boolean" => Self::Boolean,
-            _ => Self::Named(name),
-        }
-    }
-}
-
-impl From<ast::ListType> for FieldType {
-    // Spec: https://spec.graphql.org/draft/#ListType
-    fn from(list: ast::ListType) -> Self {
-        Self::List(Box::new(
-            list.ty()
-                .expect("the node Type is not optional in the spec; qed")
-                .into(),
-        ))
-    }
-}
-
-impl From<ast::NonNullType> for FieldType {
-    // Spec: https://spec.graphql.org/draft/#NonNullType
-    fn from(non_null: ast::NonNullType) -> Self {
-        if let Some(list) = non_null.list_type() {
-            Self::NonNull(Box::new(list.into()))
-        } else if let Some(named) = non_null.named_type() {
-            Self::NonNull(Box::new(named.into()))
-        } else {
-            eprintln!("{:?}", non_null);
-            eprintln!("{:?}", non_null.to_string());
-            unreachable!("either the NamedType node is provided, either the ListType node; qed")
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct Fragments {
-    map: HashMap<String, Vec<Selection>>,
-}
-
-impl From<&ast::Document> for Fragments {
-    fn from(document: &ast::Document) -> Self {
-        let map = document
-            .definitions()
-            .filter_map(|definition| match definition {
-                // Spec: https://spec.graphql.org/draft/#FragmentDefinition
-                ast::Definition::FragmentDefinition(fragment_definition) => {
-                    let name = fragment_definition
-                        .fragment_name()
-                        .expect("the node FragmentName is not optional in the spec; qed")
-                        .name()
-                        .unwrap()
-                        .text()
-                        .to_string();
-                    let selection_set = fragment_definition
-                        .selection_set()
-                        .expect("the node SelectionSet is not optional in the spec; qed");
-
-                    Some((name, selection_set.selections().map(Into::into).collect()))
-                }
-                _ => None,
-            })
-            .collect();
-        Fragments { map }
-    }
-}
-
-impl Fragments {
-    pub(crate) fn get(&self, key: impl AsRef<str>) -> Option<&[Selection]> {
-        self.map.get(key.as_ref()).map(|x| x.as_slice())
-    }
-}
 
 #[cfg(test)]
 mod tests {
